requirements.txt
requests
langfuse
python-dotenv
boto3
pandas